requirements.txt
Django==2.1
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
Faker==0.8.15
flake8==3.5.0
httplib2==0.11.3
mccabe==0.6.1
oauth2==1.9.0.post1
Pillow==5.2.0
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
text-unidecode==1.2
coverage==4.5.1
